fix(rust): clear clippy errors blocking conformance CI (esm-74r)
- Drop unused `validate` import from schema_validation.rs.
- Remove redundant `as i64` cast on searchsorted result (already i64).
- Add missing `staggering_rules: None` field in benches/benchmarks.rs
  EsmFile initializer so `--all-targets` builds again.

`cargo clippy --all-targets --all-features -- -D warnings` now passes.

+++ packages/earthsci-toolkit-rs/src/registered_functions.rs
@@ -210,7 +210,7 @@ pub fn evaluate_closed_function(
             let xs = args[1].expect_array(name, 1)?;
             Ok(ClosedValue::Integer(check_i32(
                 name,
-                searchsorted(name, x, xs)? as i64,
+                searchsorted(name, x, xs)?,
             )?))
         }
         _ => Err(ClosedFunctionError::new(
